LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-interrupted-expected.txt


FAIL interrupted navigate() promises assert_array_equals: expected property 1 to be "#1" but got "" (expected array ["", "#1", "#2"] got ["", "", "#1"])
